move tone_analyzer function to helpers.js
