+++ views/searchresults.ejs
@@ -25,6 +25,7 @@ var keyword = <%- JSON.stringify(keyword) %>;
 var url = <%- JSON.stringify(url) %>;
 var spanTags = <%- JSON.stringify(spanTags) %>;
 console.log(spanTags.fulfillmentValue);
+console.log(emotionObj);
 
 var labels = [];
 var data = [];
